Moved all Q-learning code to new file
